+++ FinancialLife/examples/dependencies.py
@@ -10,8 +10,8 @@ import os
 # third-party libraries
 
 # own libraries
-from financial_life.financing import accounts as a
-from financial_life.reports import html
+from FinancialLife.financing import accounts as a
+from FinancialLife.reports import html
 from matplotlib.pyplot import show
 
 def dependencies():