+++ FinancialLife/financing/test_status.py
@@ -8,8 +8,8 @@ from datetime import datetime, timedelta
 import unittest
 
 # own libraries
-from financial_life.financing import accounts as a
-from financial_life.financing import Status
+from FinancialLife.financing import accounts as a
+from FinancialLife.financing import Status
 
 
 class Test(unittest.TestCase):