+++ FinancialLife/test_general.py
@@ -9,8 +9,8 @@ import os
 import unittest
 
 # own libraries
-from financial_life.financing import accounts as a
-from financial_life.reports import html
+from FinancialLife.financing import accounts as a
+from FinancialLife.reports import html
 
 
 class Test(unittest.TestCase):