+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * changed some bugs for rendering HTML content
 
 # 0.9.2 (05.01.2017)
-* financial_life now supports meta-data for account objects and payments
+* FinancialLife now supports meta-data for account objects and payments
 
 # 0.9.1
 * minor issues with DataFrame support